server: move Violation class to `rule` package (#429)
Because:
- it's not a dto anymore;
- it has a field of type Rule
- it's mostly referred from that package
