+++ server/src/main/java/de/zalando/zally/rule/DefineOAuthScopesRule.kt
@@ -1,7 +1,6 @@
 package de.zalando.zally.rule
 
 import com.google.common.collect.Sets
-import de.zalando.zally.dto.Violation
 import de.zalando.zally.dto.ViolationType.MUST
 import io.swagger.models.Operation
 import io.swagger.models.Swagger
