+++ server/src/main/java/de/zalando/zally/rule/ExtensibleEnumRule.kt
@@ -1,6 +1,5 @@
 package de.zalando.zally.rule
 
-import de.zalando.zally.dto.Violation
 import de.zalando.zally.dto.ViolationType.SHOULD
 import io.swagger.models.Operation
 import io.swagger.models.Swagger
