+++ server/src/main/java/de/zalando/zally/rule/InvalidApiSchemaRule.kt
@@ -1,6 +1,5 @@
 package de.zalando.zally.rule
 
-import de.zalando.zally.dto.Violation
 import de.zalando.zally.dto.ViolationType
 import io.swagger.models.Swagger
