+++ server/src/main/java/de/zalando/zally/rule/PluralizeResourceNamesRule.kt
@@ -1,7 +1,6 @@
 package de.zalando.zally.rule
 
 import com.typesafe.config.Config
-import de.zalando.zally.dto.Violation
 import de.zalando.zally.dto.ViolationType
 import de.zalando.zally.util.PatternUtil
 import de.zalando.zally.util.WordUtil.isPlural
