+++ server/src/main/java/de/zalando/zally/rule/SuccessResponseAsJsonObjectRule.kt
@@ -1,6 +1,5 @@
 package de.zalando.zally.rule
 
-import de.zalando.zally.dto.Violation
 import de.zalando.zally.dto.ViolationType
 import io.swagger.models.ComposedModel
 import io.swagger.models.ModelImpl
