+++ server/src/main/java/de/zalando/zally/rule/UsePasswordFlowWithOauth2Rule.kt
@@ -1,6 +1,5 @@
 package de.zalando.zally.rule
 
-import de.zalando.zally.dto.Violation
 import de.zalando.zally.dto.ViolationType
 import io.swagger.models.Swagger
 import io.swagger.models.auth.OAuth2Definition
