+++ server/src/test/java/de/zalando/zally/rule/ExtractBasePathRuleTest.kt
@@ -1,6 +1,5 @@
 package de.zalando.zally.rule
 
-import de.zalando.zally.dto.Violation
 import de.zalando.zally.dto.ViolationType
 import de.zalando.zally.getFixture
 import de.zalando.zally.swaggerWithPaths
