+++ server/src/test/java/de/zalando/zally/rule/NoProtocolInHostRuleTest.kt
@@ -1,6 +1,5 @@
 package de.zalando.zally.rule
 
-import de.zalando.zally.dto.Violation
 import de.zalando.zally.getFixture
 import io.swagger.models.Swagger
 import org.assertj.core.api.Assertions.assertThat
