+++ server/src/test/java/de/zalando/zally/statistic/RestReviewStatisticsTest.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.databind.JsonNode;
 import de.zalando.zally.apireview.ApiReview;
 import de.zalando.zally.apireview.RestApiBaseTest;
 import de.zalando.zally.dto.ApiDefinitionRequest;
-import de.zalando.zally.dto.Violation;
 import de.zalando.zally.dto.ViolationType;
 import de.zalando.zally.rule.InvalidApiSchemaRule;
+import de.zalando.zally.rule.Violation;
 import de.zalando.zally.util.TestDateUtil;
 import org.junit.Test;
 import org.springframework.http.HttpStatus;
